(refactor) LineGenerator 삭제 -> Generator 인터페이스로 통합
