+++ src/main/java/ladder/generator/Generator.java
@@ -2,9 +2,11 @@ package ladder.generator;
 
 import ladder.domain.ladder.Ladder;
 import ladder.domain.ladder.LadderHeight;
+import ladder.domain.ladder.Line;
 
-public interface LadderGenerator extends Generator {
+public interface Generator {
 
     Ladder generate(int countOfPlayer, LadderHeight height);
+    Line generate(int countOfPlayer);
 
 }
