--- src/main/java/ladder/generator/LineGenerator.java
@@ -1,10 +1,0 @@
-package ladder.generator;
-
-import ladder.domain.ladder.Line;
-
-@FunctionalInterface
-public interface LineGenerator {
-
-    Line generate(int countOfPlayer);
-
-}
