+++ src/test/java/ladder/domain/ladder/LineTest.java
@@ -1,7 +1,7 @@
 package ladder.domain.ladder;
 
-import ladder.generator.DefaultLineGenerator;
-import ladder.generator.LineGenerator;
+import ladder.generator.DefaultLadderGenerator;
+import ladder.generator.Generator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -14,16 +14,16 @@ class LineTest {
 
     @BeforeEach
     void init() {
-        LineGenerator lineGenerator = new DefaultLineGenerator(() -> true);
-        line = lineGenerator.generate(4);
+        Generator ladderGenerator = new DefaultLadderGenerator(() -> true);
+        line = ladderGenerator.generate(4);
     }
 
     @Test
     @DisplayName("Generator 로 line 생성")
     void generateLine() {
-        LineGenerator lineGenerator = new DefaultLineGenerator(() -> true);
-        Line line = lineGenerator.generate(4);
-        assertThat(line).isEqualTo(lineGenerator.generate(4));
+        Generator ladderGenerator = new DefaultLadderGenerator(() -> true);
+        Line line = ladderGenerator.generate(4);
+        assertThat(line).isEqualTo(ladderGenerator.generate(4));
         assertThat(line.getPoints()).size().isEqualTo(4);
     }
 
@@ -42,8 +42,8 @@ class LineTest {
     @Test
     @DisplayName("라인에서 move(위치값) 호출 -> 위치값 계산 후 반환(정지)")
     void notMove() {
-        LineGenerator lineGenerator = new DefaultLineGenerator(() -> false);
-        Line line = lineGenerator.generate(4);
+        Generator ladderGenerator = new DefaultLadderGenerator(() -> false);
+        Line line = ladderGenerator.generate(4);
         assertThat(line.move(1)).isEqualTo(1);
     }
